device-telemetry-alerts/services/worker/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.8
prometheus-client==0.20.0
boto3==1.34.162
